[STEP2] Controller, View 수정
