+++ src/main/kotlin/lotto/view/LottoController.kt
@@ -1,14 +1,13 @@
 package lotto.view
 
 import lotto.domain.LottoShop
-import lotto.domain.LottoStatistic
 
 fun main() {
-    val money = LottoInputController.inputMoney()
+    val money = LottoInputView.inputMoney()
     val lottos = LottoShop.sellByMoney(money)
-    LottoOutputController.printLottos(lottos)
+    LottoOutputView.printLottos(lottos)
 
-    val winningNumbers = LottoInputController.inputWinningLottoNumbers()
-    val lottoStatistic = LottoStatistic(winningNumbers = winningNumbers, lottos = lottos)
-    LottoOutputController.printLottoStatistic(lottoStatistic)
+    val winningNumbers = LottoInputView.inputWinningLottoNumbers()
+    val lottosResult = lottos.calculateResults(winningNumbers)
+    LottoOutputView.printLottoResults(lottosResult)
 }
